requirements.txt
streamlit
ultralytics
opencv-python
Pillow
numpy
pyyaml
